Major script overhaul. Laying the foundation for future code.
creating a script naming convention to make the code much easier to navigate. Added a detailed recipe section for more complicated filtering and recipes based on specific ingredient amounts. added code for the inventory and detailed inventory to make it easier for minne to fill out with recipes. Also added a new json file for detailed recipes